Allow undo actions to be applied on rollbacks and discard unwanted undo.
We always perform rollback actions after cleaning up the current
(sub)transaction.  This will ensure that we perform the actions immediately
after an error (and release the locks) rather than when the user issues
Rollback command at some later point of time.  We are releasing the locks
after the undo actions are applied.  The reason to delay lock release is
that if we release locks before applying undo actions, then the parallel
session can acquire the lock before us which can lead to deadlock.

To improve the efficiency of the rollbacks, we create three queues and a
hash table for the rollback requests.  A Xid based priority queue which
will allow us to process the requests of older transactions and help us to
move oldesdXidHavingUndo (this is a xid-horizon below which all the
transactions are visible) forward.  A size-based queue which will help us
to perform the rollbacks of larger aborts in a timely fashion so that we
don't get stuck while processing them during discard of the logs.  An error
queue to hold the requests for transactions that failed to apply its undo.
The rollback hash table is used to avoid duplicate undo requests by
backends and discard worker.

Undo launcher is responsible for launching the workers iff there is some
work available in one of the work queues and there are more workers
available.  The worker is launched to handle requests for a particular
database.

The discard worker is responsible for discarding the undo log of
transactions that are committed and all-visible or are rolled-back.  It
also registers the request for aborted transactions in the work queues.
It iterates through all the active logs one-by-one and tries to discard the
transactions that are old enough to matter.

Designed by: Andres Freund, Amit Kapila, Robert Haas, and Thomas Munro
Author: Amit Kapila, Dilip Kumar, Kuntal Ghosh, and Thomas Munro